My Second Commit
Number type game  chooses  a random number between 1 and 100, then challenges the player to guess the number in 10 turns. After each turn the player is  told if they are right or wrong, and if they are wrong, the player is also told whether the guess was too low or too high. It also tells the player what numbers they previously guessed. The game will end once the player guesses correctly, or once they run out of turns. When the game ends, the player is  given an option to start playing again.